[portability] Directly use int64_t instead of off_t.

+++ src/default_blockinput.c
@@ -49,7 +49,7 @@
 #endif
 
 #ifdef _WIN32
-static ssize_t pread(int fd, void *buf, size_t count, off_t offset)
+static ssize_t pread(int fd, void *buf, size_t count, int64_t offset)
 {
     OVERLAPPED ov;
     DWORD      got;
@@ -95,7 +95,7 @@ static int _def_close(udfread_block_input *p_gen)
 static uint32_t _def_size(udfread_block_input *p_gen)
 {
     default_block_input *p = (default_block_input *)p_gen;
-    off_t pos;
+    int64_t pos;
 
     pos = lseek(p->fd, 0, SEEK_END);
     if (pos < 0) {
@@ -105,17 +105,17 @@ static uint32_t _def_size(udfread_block_input *p_gen)
     return (uint32_t)(pos / UDF_BLOCK_SIZE);
 }
 
-static int _def_read(udfread_block_input *p_gen, uint32_t lba, void *buf, uint32_t nblocks, int flags)
+static int _def_read(udfread_block_input *p_gen, uint64_t lba, void *buf, uint32_t nblocks, int flags)
 {
     (void)flags;
     default_block_input *p = (default_block_input *)p_gen;
 
     size_t bytes, got;
-    off_t  pos;
+    int64_t  pos;
 
     bytes = (size_t)nblocks * UDF_BLOCK_SIZE;
     got   = 0;
-    pos   = (off_t)lba * UDF_BLOCK_SIZE;
+    pos   = (int64_t)lba * UDF_BLOCK_SIZE;
 
     while (got < bytes) {
         ssize_t ret = pread(p->fd, ((char*)buf) + got, bytes - got, pos + got);
